refactor: EventCell 분리
