+++ src/components/calendarView/MonthView.tsx
@@ -13,6 +13,7 @@ import {
   VStack,
 } from '@chakra-ui/react';
 
+import { EventCell } from './EventCell';
 import { WEEKDAYS } from '../../constants/constants';
 import { Event } from '../../types';
 import { formatDate, formatMonth, getEventsForDay, getWeeksAtMonth } from '../../utils/dateUtils';
@@ -64,25 +65,7 @@ export const MonthView = ({ currentDate, holidays, filteredEvents, notifiedEvent
                           </Text>
                         )}
                         {getEventsForDay(filteredEvents, day).map((event) => {
-                          const isNotified = notifiedEvents.includes(event.id);
-                          return (
-                            <Box
-                              key={event.id}
-                              p={1}
-                              my={1}
-                              bg={isNotified ? 'red.100' : 'gray.100'}
-                              borderRadius="md"
-                              fontWeight={isNotified ? 'bold' : 'normal'}
-                              color={isNotified ? 'red.500' : 'inherit'}
-                            >
-                              <HStack spacing={1}>
-                                {isNotified && <BellIcon />}
-                                <Text fontSize="sm" noOfLines={1}>
-                                  {event.title}
-                                </Text>
-                              </HStack>
-                            </Box>
-                          );
+                          return <EventCell event={event} notifiedEvents={notifiedEvents} />;
                         })}
                       </>
                     )}
